+++ apps/openmw/mwgui/recharge.cpp
@@ -119,7 +119,7 @@ void Recharge::updateView()
 
         Widgets::MWDynamicStatPtr chargeWidget = mView->createWidget<Widgets::MWDynamicStat>
                 ("MW_ChargeBar", MyGUI::IntCoord(72, currentY+2, 199, 20), MyGUI::Align::Default);
-        chargeWidget->setValue(iter->getCellRef().getEnchantmentCharge(), enchantment->mData.mCharge);
+        chargeWidget->setValue(static_cast<int>(iter->getCellRef().getEnchantmentCharge()), enchantment->mData.mCharge);
         chargeWidget->setNeedMouseFocus(false);
 
         currentY += 32 + 4;
@@ -149,11 +149,11 @@ void Recharge::onItemClicked(MyGUI::Widget *sender)
     MWMechanics::CreatureStats& stats = player.getClass().getCreatureStats(player);
     MWMechanics::NpcStats& npcStats = player.getClass().getNpcStats(player);
 
-    float luckTerm = 0.1 * stats.getAttribute(ESM::Attribute::Luck).getModified();
+    float luckTerm = 0.1f * stats.getAttribute(ESM::Attribute::Luck).getModified();
     if (luckTerm < 1|| luckTerm > 10)
         luckTerm = 1;
 
-    float intelligenceTerm = 0.2 * stats.getAttribute(ESM::Attribute::Intelligence).getModified();
+    float intelligenceTerm = 0.2f * stats.getAttribute(ESM::Attribute::Intelligence).getModified();
 
     if (intelligenceTerm > 20)
         intelligenceTerm = 20;
@@ -161,7 +161,7 @@ void Recharge::onItemClicked(MyGUI::Widget *sender)
         intelligenceTerm = 1;
 
     float x = (npcStats.getSkill(ESM::Skill::Enchant).getModified() + intelligenceTerm + luckTerm) * stats.getFatigueTerm();
-    int roll = std::rand()/ (static_cast<double> (RAND_MAX) + 1) * 100; // [0, 99]
+    int roll = static_cast<int>(std::rand() / (static_cast<double> (RAND_MAX)+1) * 100); // [0, 99]
     if (roll < x)
     {
         std::string soul = gem.getCellRef().getSoul();
@@ -197,10 +197,10 @@ void Recharge::onItemClicked(MyGUI::Widget *sender)
 
 void Recharge::onMouseWheel(MyGUI::Widget* _sender, int _rel)
 {
-    if (mView->getViewOffset().top + _rel*0.3 > 0)
+    if (mView->getViewOffset().top + _rel*0.3f > 0)
         mView->setViewOffset(MyGUI::IntPoint(0, 0));
     else
-        mView->setViewOffset(MyGUI::IntPoint(0, mView->getViewOffset().top + _rel*0.3));
+        mView->setViewOffset(MyGUI::IntPoint(0, static_cast<int>(mView->getViewOffset().top + _rel*0.3f)));
 }
 
 }
